services/requirements.txt
fastapi
confluent_kafka
uvicorn
python-dotenv
pymongo
httpx
beautifulsoup4
lxml
motor